Backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
dnspython
python-dotenv
python-jose[cryptography]
bcrypt
python-multipart
httpx
